format structure
